Use SetUpWithNewChrome fo full restore test
Instead of starting directly Chrome, the newly added
browserfixt.SetUpWithNewChrome is now helpful for selecting which type
of browser to be used e.g. Lacros or Ash Chrome. There are no functional
changes in this CL.

A possible follow-up work is to give support for Lacros full restore in
this same test. This can be used to stress the lacrosfixt execution
path, get the Lacros full restore support working correct, which among
other things is helpful for Lacros performance measurements (See
https://crrev.com/c/3508772).

BUG=None
TEST=tast run device apps.FullRestoreAlwaysRestore

Change-Id: I615a37716943213e30ae3514abed8cb947f4c519