tast-tests: Add wifi.CheckIntelFWDump and two more to wifi_matfunc
Three tests are to be added to wifi_matfunc: wifi.CheckIntelFWDump
wifi.HostapHwsim.validity and wifi.UdevRename. They are  are running as
part of bvt-tast-cq but not in wifi_matfunc. That makes it much easier
to have an at-a-glance view of the health of our tests by looking at
wifi_matfunc results. It's also better for our partners who are only
required to run wifi_matfunc.

The "group:wificell" is required to run wifi_matfunc as pointed by
control file wificell-func-stable. So the next CL will make above 3
tests require wificell fixtures.

BUG=b:226981380
TEST=run test_that --board=octopus $DUT suite:wifi_matfunc and then \
     look into logs to confirm that those 3 tests are added

Change-Id: Ief1480c070210d3fd675245b73740770a65c89c8