health: Wait for DRM resource initialized
crosHealthdRunning fixture restarts the ui. And drm resources need a
little time to initialize. So cros_healthd may retrieve empty info while
modetest command may retrieve non-empty info. This causes unstable.

An approach is to poll the connector status. If we can see the encoder
id, it means the initialization is done, so we can move on. If after 10
seconds, we still can't find the encoder id, it might be in hibernate
state, we just left a log and run the rest of the test and make sure
cros_healthd reports nothing.

BUG=b:225766968
TEST=(bob device) tast run $DUT_IP health.ProbeDisplayInfo

Change-Id: I80eb2e4ba144337408e38f088da62a37809f15c4