dlp: Wait for quiescence in Clipboard, History, MixedBrowserType, Dnd.
Some DLP Tast tests were failing randomly if the internet connection to
the DUT was slow. In this case, the Tast test attempts to perform an
an action (e.g. copy data) from a not fully loaded website. In this CL,
we modify the Clipboard, ClipboardHistory, ClipboardMixedTypeBrowsers,
and DragDrop tests by adding a WaitForQuiescence() call before
performing an action on a website if there wasn't already
a WaitUntilExists() call.

BUG=chromium:1304833, chromium:1304816, chromium:1304827,
chromium:1290248
TEST=tast <ip> dlp.DataLeakPreventionRulesList*

Change-Id: I9704c15324516b5fa5eec194a4d14f80bd570bc5