[Tast] Verify app installation with launcher sorted
This test verifies app installation when app list is sorted. To ensure
that the app's name and icon are fixed, a mock android app is built by
this CL.

The test procedure:

(1) Sort app list with alphabetical order or color order.
(2) Install the mock app when app list is under temporary sort.
(3) Uninstall the mock app.
(4) Install the mock app again to verify app installation behavior under
permanent sort.

To ensure that the mock app has the fixed app name and icon, this CL
creates a mock android app for testing.

BUG=chromium:1227736
TEST=launcher.InstallAppWithAppListSorted.*

Change-Id: I0f0430cf6f42fda832430665f21b9a77ecd3ade6