tast-tests: Compute package install time dynamically
Currently the test waits as long as the test timeout to wait for
packages to install and the session has locked by then so the screenshot
is not useful.

In this change, we're computing the timeout to wait for packages to
install based on no. of packages being installed so we can timeout
early.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: Ie294fd491266ffeeb2c5288a229e6684212982cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3566278
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Auto-Submit: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>