lacros: Remove thread priority code
Wayland event watcher thread is gone since the introduction of major
changes in the event loop code (crrev.com/c/3386963) and the Wayland
thread priority flag became unavailable (crrev.com/c/3566134). Therefore
this removes Wayland flag and its Lacros fixture.

This CL also removes Lacros fixture related to thread priorities
testing, and in particular its only usage in DocsCUJ benchmark. I
believe we can all agree now that it was a bad idea to use one benchmark
only as a test bed for analysing all the effects of thread priority and
scheduling related in Lacros. It is a work that requires a larger
benchmark scope and, as such, it is being carried now more carefully.
See https://crrev.com/c/3498696.

BUG=chromium:1262133
TEST=tast run $device lacros.DocsCUJ*

Change-Id: Iaddc11a06fd583ac5cd7f7b02b330fb661aa893b