unicorn login tast: verify unicorn user logged into browser
Update the unicorn login tast test to verify that the unicorn
user is logged into the browser. A todo still remains to do this
check for both lacros and ash browsers, but this is blocked on
go/lacros-tast-simplification, which will allow this test to use
lacros without completely rewriting the fixture.

BUG=chromium:1218629
TEST=Ran test on VM and device

Change-Id: Ia37cce027243055e2839a1da77836f3964cabfc9