network: ARCVPNConnect: Add basic connection tests
Add some basic test functionality around mirroring host VPNs in ARC.
- ArcHostVpnService isn't launched if the flag isn't enabled
- ArcHostVpnService does start when the host VPN starts (w/ flag enabled)
- VPN is reachable on the host and ARC with ArcHostVpnService
- Unexpectedly stopping ArcHostVpnService doesn't affect VPN
  reachability in ARC
- Disconnecting the host VPN also stops ArcHostVpnService

Note that ARCVPNConnect doesn't exhaustively test different host VPNs.
The VPN chosen is unimportant for ARC functionality, but to issue the
ADB commands in the test, the user needs to be logged in on the dut. In
the test setup, AuthTypeCert VPNs will log the user out while setting up
the cert store so we avoid those.

BUG=b:147256449
TEST=tast run $DUT network.ARCVPNConnect*

Change-Id: I7669a942e7ce46840104c5e7e7fd322e2b53a174