hwsec: Glob in tpm_clearer.go don't work
rm is being called with the actual entry in SystemStateFiles.
The globs used are not evaluated and we never remove the files.

This makes enrollment without reboot fail if install_attributes.pb
was present before.

install_attributes.pb is the only file in /run/lockbox/.
/var/lib/devicesettings/ can contain any number of policy files,
use a glob to remove them.

BUG=b:223565358
TEST=tast run ip policy.AllowDinosaurEasterEgg policy.AllowDinosaurEasterEggEnrolled

Change-Id: I64ff849e150a67ea048fb0c0b8c7a3ff67ca47fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3569073
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>