faft: Find base gpios via ectool
Add one function to be implemented with the host
command 'ectool', and to look for gpios that are
relevant to control on a detachable base. This
function would be useful for a new test on checking
the notification window upon base ec update.

BUG=b:200305434
TEST=manual

Change-Id: I84bb42e3da8ddf8d340a20b73ce2d9ac7ee08ea7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3574185
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Arthur Chuang <arthur.chuang@cienet.com>
Tested-by: Arthur Chuang <arthur.chuang@cienet.com>