tast-tests: repeat active scan if no probe request is captured in the pcap file
Probe requests from the client may not be captured in the pcap file
if the AP keeps sending beacon frames. To prevent this from failing
the test, repeat the active scan until the probe request is captured
in the pcap file, or it reaches the max number of scans.

BUG=b:225073589
TEST=tast run -var=router=localhost:2210 -var=pcap=localhost:2220
localhost:2200  wifi.ChannelScanDwellTime

Change-Id: Ie8f303ad8a75d71971161bcb799e812b0f0fafd7