sandboxed_services: Print baseline/exclusion list entry on failures.
These failures often correspond to short-lived root processes that are
only present on specific boards. The failures can be hard to fix
because we don't always have access to the failing boards.

However, the failures are normally easy to diagnose by looking up the
name of the process, and checking that it is indeed a short-lived
script or tool. By printing the expected baseline/exclusion entry we
make it easier for ourselves to fix some failures, especially
considering that process names are truncated by the kernel.

There is little risk of this resulting in folks adding baselines for
processes that shouldn't have one, since CLs require domain and Tast
reviews.

BUG=b:228320883
TEST=Run a process as root: yes > /dev/null
TEST=Run the test, see the error message:
TEST=A baseline entry for this process would look like:
TEST={"yes", "root", "root", 0}
TEST=Copy and paste the baseline mechanically, the test passes.

Change-Id: I00da68014a0542a6be73e2cc79d47662afc47a34