wifi: Add a flexible hostapd configuration for Passpoint
This CL refactors the wiredhostapd package to allow a versatile
configuration (depending on the usage) while keeping the current server
wrapper. The hostapd package proposes a server implementation with an
interface for the configuration. Each server user can implement the
interface to provide its own specific configuration. Two configurations
are already available: a wired radius server on one hand, and a
Passpoint access point on the other hand.

BUG=b:162258594
TEST=tast run ${DUT} network.Ethernet8021X.*

Change-Id: I1f2031f0f35d46d0bdc8e52d95bccb66b7966287