hwsec: Switch to using password in WebauthnU2fMode
Most flakiness observed in tastboard for this test is about failing to
set up a PIN in settings page. I doubt that it is about some lacking
policy control in this test, but that doesn't really matter because we
already have a WebauthnUsingPIN test for monitoring PIN WebAuthn. This
test's purpose is to monitor behavior of UP/UV path is correct and now
we have a less flaky way of performing WebAuthn, that is using password.
Switch to use password for this test.

BUG=b:228571560
TEST=tast run -var=servo=$SERVO $DUT hwsec.WebauthnU2fMode

Change-Id: Iccd34f019c1d7a17f88cfb60b19bf1ad28d78b01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3578594
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>