autoupdate: Add clarification to error message
The error message that appears when the TLW service is not running
does not seem to be informative enough. Add small clarification to
the error message to easily identify that the issue can be that the
TLW service is not running and it needs to.

BUG=None
TEST=tast run -var=updateutil.tlwAddress="localhost:7151" <DUT> \
autoupdate.BasicNToN shows error message with clarification when
TLW is not running

Change-Id: Ib1b0c13005eb95248bfdb1a12e8a49ec98ec043e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3579323
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>