familylink: Verify child users logged into browser
Update the unicorn and geller login tast test to verify that the
user is logged into the browser and recognized as a child. A todo
still remains to do this check for both lacros and ash browsers,
but this is blocked on go/lacros-tast-simplification, which will
allow this test to use lacros without completely rewriting the fixture.

BUG=chromium:1313098
TEST=Ran test on VM and device

Change-Id: Ic4fcee85bfa3a24954d2a1a4838f7d450eb5b8f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3579521
Reviewed-by: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Tested-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Galen Corey <galenemco@chromium.org>