camera: Make tests compatible after moving grid/timer out of settings UI
According to the new Settings UI design, the grid/timer option will be
moved out of settings UI and be placed at options group. Ideally, the
tests should be moved to somewhere like CCAUIPreviewOptions.

This CL also tweaks the test logic for mirror button since according to
the new UI spec, it also has a option panel to be consistent with
grid/timer options.

BUG=b:215484798
TEST=tast run [DUT] camera.CCAUISettings camera.CCAUIPreviewOptions

Change-Id: Ie33b7543b55916b7d23e8f2e87c774b0d48781de