tast-tests: Close streams to get WebRTC data in ui.MeetCUJ.*
tast.ui.MeetCUJ.* are intended to collect some WebRTC histograms that
tools/metrics/histograms/metadata/web_rtc/histograms.xml in the chromium
code base describes as "Recorded when a stream is removed." So the tests
will remove the video streams before expecting to get data for these.
Also, one of them is sought with the wrong name: the tests expect
WebRTC.Video.DroppedFrames.RateLimiter (with an upper case L in Limiter)
but it is actually called WebRTC.Video.DroppedFrames.Ratelimiter (with a
lower case l in limiter).

BUG=b:228863125
TEST=tast run <DUT> ui.MeetCUJ.16p

Change-Id: Ia98ff6abe577c5b0bad9f90e76a8916847d1f847