Fix OS finch smoke test
The test started failing due to the OobeConsolidatedConsent feature
being enabled by default. This feature's interaction with the
OobeAPI-based Tast login causes the variations_compressed_seed and
variations_seed_signature fields to not appear in Local State. As a
quick-fix to get the test working again, just disable the feature. I
will follow up and see what changes need to be made to the OOBE test API
to get it working in the default state (maybe some metrics opt-in is
being skipped in the enabled case)

BUG=b:226500971
TEST=tast run <dut> ui.ChromeVariationsSmoke

Change-Id: I1391e556210ba530513d8032dc04ef8cc2373860