tast-tests: retry on package list failure due to Chrome crash
Some times Chrome process crashes during managed provisioning due to a
corrupted extension being reloaded. See b/229745779

In this change, we'll retry the test in the stable variant to verify the
provisioning flow if we detect that Chrome process has crashed. In the
unstable variant, the test will continue to fail and we'll not loose
sight of this problem (and other similar problems in the future).

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: I9f1df912fb085f80964e092ed4156ff3b291dd47