video: update switch_1080p_720...hevc.json external file
Said file was generated with a buggy implementation. After fixing
the implementation, now the test fails. This CL updates the
external file to point to a corrected md5sum json, which was
generated using the previous CL updated script and

~/trunk/src/platform/tast-tests/tools/generate_external_file.py switch_1080p_720p_240frames_20220411.hevc.json video

(Had to edit minimally the generated resource suffix though).

BUG=b:227247655
BUG=b:227238318
TEST=Ideally video.DecodeAccelVD.hevc_resolution_switch* on volteer
but this is pending me having access to such a device.

Change-Id: If838fcbfb2a280c0e2446a53bc9e23a3b3a279ae