Fix failing test: AppGeditFilesharing
AppGeditFilesharing test has been failing because it used Text
ChromeOS app, which was not always installed in crostiniBusterLargeContainer.
The absence of Text app could be a bug that will be investigated
separately. However, AppGeditFilesharing test does not need to depend on
this app.

Adjust the test to leverage Chrome instead of Test app for viewing the files.

BUG=b:226240873
TEST=tast run dut crostini.AppGeditFilesharing

Change-Id: Ie5c97fb8943ba5ba5ed644d7b5227d7e93cd3a05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3581750
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>