health: Split progif in another sub test
Instead of setting volteer2 as the hwdeps, split the progif checking
into a sub test. This is because the volteer tast could pass if it use
volteer2 to run the tast. It caused in unstable CQ.
Split the checking of progif into another test and set it informational.
The rest of the test didn't change and should be stable as we have run
them for months.

BUG=b:199683963
TEST=tast run $DUT health.ProbeBusInfo
TEST=tast run $DUT health.ProbeBusInfo.progif

Change-Id: I8577dc16715e92de8551cf911be4a8a395cb67b2