tast-tests: Split launcher.BubbleSmoke into two tests
Simplify launcher.BubbleSmoke so it just tests opening and closing the
launcher with mouse and keyboard. This will make it easier to stabilize
this test and promote it to critical.

Extract the part of the test that launches the OS Settings app to a new
test launcher.BubbleLaunchApp. This part of the test has been flaky,
so keep it informational for now.

BUG=chromium:1289371
TEST=tast run <dut> launcher.Bubble*

Change-Id: Iab1aa854de7a67f41363feffca1a971fe9def31c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3582821
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: James Cook <jamescook@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>