u2fd: Promote WebauthnUsingPassword.gsc to CQ
Remove the informational attribute from the u2fd.WebauthnUsingPassword
test (gsc) so it runs on CQ. The tpm1 version is already run on CQ, and
this version isn't because there's some flakiness at that time. Now this
test has become more stable: only one failure of known issue and one
flaky failure in last ~30 versions shown on tastboard. Sometimes there's
no TPM 1.2 builders run in CQ so it's important both versions of this
test is in, so that if anything breaks in WebAuthn we can catch it.

BUG=None
TEST=tastboard consistently passing

Change-Id: Ib990bfc20a44a0a8006f589adea191ae36c0aced
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3583625
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>