cuj: Use isLacros to determine whether to open the lacros browser
In cuj/ui.go, the GetBrowserStartTime() function uses argument
lFixtVal. It was used in M101 in the lacros.Connect() to connect to UI
opened lacros-Chrome. The lacros.Connect() was changed in M102 and we
simplify the GetBrowserStartTime() signature by just passing in the
isLacros bool.

BUG=b:227525974
TEST=fast_build.sh -T

Change-Id: I467066d891f17eace253f9a8b47981aabbe30072
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3585275
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Xibin Liu <xibin@google.com>
Commit-Queue: Xibin Liu <xibin@google.com>