refactor: Support Enable/DisableFeatures args for Lacros
Lacros features and additional args used to launch lacros-chrome should
be delimited by '####' and passed in from ash-chrome as a single arg
with --lacros-chrome-additional-args.
Example:
  --lacros-chrome-additional-args="--enable-features=Feature1,Feature2####--disable-features=Feature3####--foo=bar"
will result in multiple arguments passed to lacros-chrome:
  --enable-features=Feature1,Feature2
  --disable-features=Feature3
  --foo=bar

In the past if --disable-features is set twice the only last one counts
and all others get ignored, which is not expected.

BUG=chromium:1302531
TEST=Run the tests with the lacros features enabled or disabled
TEST=  tast run policy.HistoryClustersVisible.lacros
TEST=  tast run lacros.GpuCUJ.maximized_delegated
TEST=Run regression tests
TEST=  tast run '("group:mainline" && "dep:lacros" && !informational)'

Change-Id: I47b2a2065e9576f358ecf12c5450f6f45dded951