Fix CheckSkuPurchase test
Changes in crrev.com/c/3581523 introduced a javascript call that assumes
a presence of the element on the page. However, in CheckSkuPurchase test
the presence of the element is awaited in waitForStableElementByJs, which
is invoked after the introduced code. As result, CheckSkuPurchase started
to fail, because the element is not yet available.

Adjust the tests to wait for the element presence before interacting
with it.

BUG=b:228783666
TEST=tast run dut appsplatform.CheckSkuPurchase.vm

Change-Id: Ic683b29ca05cd32abc2080f1d30cffdcde93e3b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3586566
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>