Tast test: Fix NotificationClosePerf.one_at_a_time_arc.
The quick settings might get close while we are interacting with the ARC
testing app, so we have to reopen it whenever we delete an ARC
notification through this app. Also, in some devices the entire test
took a long time, so we have to increase the timeout value.

BUG=chromium:1302447
TEST=tast run <IP> ui.NotificationClosePerf.one_at_a_time_arc

Change-Id: I20ce06f5771c3de0b37444d3cf812d0e96299a81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3584854
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Andre Le <leandre@chromium.org>
Commit-Queue: Andre Le <leandre@chromium.org>