tast-tests: openwrt support for IfaceManipulation
This updates the OpenWrt Router implementation with support for the
IfaceManipulation interface. Since the implementation of "ip" on
OpenWrt does not support the brief output format, the ip runner has
also been set to not have this support so the full output format will
be used instead. This was necessary in order to support the calls
to "ip" used in the functions of IfaceManipulation.

BUG=b:211517577
TEST=tast run -var=router=$OPENWRT_ROUTER -var=pcap=$LEGACY_PCAP $DUT \
wifi.MissingBeacons

Change-Id: I75fd88a91543d6a6d1297ef4a8c11af72145d4cc