tast-tests: ip showLink still possible even if -br format not supported
This allows hosts that have ip implementations without -br support to
still be used with the ip runner (specifically OpenWrt). Full ip -br
support, specified with the new "BriefLinkShow" feature toggle, is the
default. When the "BriefLinkShow" feature toggle is set to false, the
-br flag will not be used and the full ip format will be parsed
instead of the brief format. New unit tests have been added to parse
this new format as well.

BUG=b:211517577
TEST=tast unit tests passed

Change-Id: I95e83edda288eec7dfefe925ba725f70d2b27d48