mta: Fix taskmanager tests
This CL fixes the following tests:
* taskmanager.BrowserEndProcessDisabled
* taskmanager.CloseTab
* taskmanager.DefaultUIAndFunctionality
* taskmanager.DisplayAllTabs
* taskmanager.EndProcess

The tests above have following issues:
1. The Process node in the Task manager failed to get focused.
This CL fixes this issue by re-design the way to select a process node.

2. The shortcut Ctrl+T didn't work as expected to create a new browser tab.
This CL fixes this issue by using (*chrome.Chrome).NewConn() instead.

3. The title of some websites are not always static and might cause tests to be failed.
This CL fixes this issue by using the chrome.tab API to acquire the correct title.

4. Plugin node didn't show up on the web page in time.
This CL fixes this issue by increasing the timeout to stabilize the procedures.

5. The website Zoom takes too long to load and causes tests to be failed.
This CL fixes this issue by replacing Zoom with Twitter.

6. The name of the YouTube plugin in the Task manager has changed.
This CL fixes this issue by updating node's name of a plugin process.

7. taskmanager.CloseTab fails on closing a web page tab under tablet mode due to the close button not being visible.
This CL fixes the issue by forcing the test to run under clamshell mode.

BUG=b:230567638, b:241943434
TEST=tast unit test passed
TEST=tast run <DUT> taskmanager.BrowserEndProcessDisabled
TEST=tast run <DUT> taskmanager.CloseTab
TEST=tast run <DUT> taskmanager.DefaultUIAndFunctionality
TEST=tast run <DUT> taskmanager.DisplayAllTabs
TEST=tast run <DUT> taskmanager.EndProcess

Change-Id: I7a1d494fc90d0120dc6eaa1312f6a02705a4754c