Tast: ensure faillog copy doesn't overwrite existing logs
When a remote test fails, `testHookRemote` uses the `Faillog` service to
get some logs, then copies them to the host. The copying is done with
`linuxssh.GetFile`, which deletes the destination if it already
exists. This has the unfortunate result of deleting any faillogs
directly created by a remote service, e.g. from calling
`faillog.DumpUITree`.

Fix by putting the faillog results in a timestamped subdirectory.

BUG=b:229284804
TEST=Call `faillog.DumpUITree` from a remote service in a remote test that is failing.
TEST=Inspect the test results under `/tmp/tast/results/latest/tests/install.OsInstall/faillog`
TEST=Verify that `ui_tree.txt` is present as well as a timestamped subdirectory like `20220415-141604.042784` containing additional files.

Change-Id: Ie915a57ff72f3183ca179bad7dd62fb1220b9d85