arcappgameperf: Add fixture implementation for TAPE.
Implements the fixture based implementation for TAPE which requests an
account from the endpoint, stores the username on the DUT, and provides
convenience methods for associating the leased account with the password
from the private credentials. Additionally, implements a leased account
fixture which can be used for Roblox tests (which now live in the
private repository).

BUG=b:222311973
TEST=fast_build.sh -T

Change-Id: I42e31a0281b5abdaca012bea23b23c47970b51c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3591872
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Tested-by: David Welling <davidwelling@google.com>
Commit-Queue: David Welling <davidwelling@google.com>