tast-tests: Add lacros variant to firmware.ECVerifyVK
The test needs Lacros enabled to verify that the virtual keyboard
is present in the tablet mode.

This also includes testability improvement to share the chrome instance
between two services
and a fix to update the query of finding the visible VK.

See go/lacros-tast-porting for the porting guidelines.

BUG=b:228326787
TEST=Start manually servod from the labstation when needed
TEST=  $ ssh $SERVO_HOSTNAME
TEST=  localhost ~ # start servod PORT=9999
TEST=tast run -var=servo=localhost:9999:ssh:<LOCAL_SERVO_SSH_PORT> ${DUT} \
TEST=  firmware.ECVerifyVK.lacros \
TEST=  firmware.ECVerifyVK*

Change-Id: Id2d14d0bb02443e5adee699e897f04f3945c3fff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3589131
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Jonathan Fan <jonfan@chromium.org>
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>