Exclude Ultima model from SocProperties tast
Ultima will stay on P container (second tab in go/arcvm-migration). So
stop running the arc.SocProperties tast test on them.

It would be "easy" to fix with a regex tweak, as the only difference is
the CPU model name has extra spaces around it [0], but I'd rather be
conservative in order to avoid combinatorially increasing complexity in
case we expand the conditional parts of the parsing regexes.

[0] |Intel(R) Celeron(R) CPU  N3150  @ 1.60GHz|

BUG=b:175610620
TEST=arc.SocProperties (recompile only)

Change-Id: I737dfb0ade8b7ee6396128d41a1e136bde6e08bc