Disable fwupd in power tests
Power tests are flaky because the standard power test setup code is
noticing that powerd is restarting during tests. Powerd is restarting
because fwupd is starting. Fwupd is starting because Chrome is asking it
to check.

This test stops Chrome from starting fwupd by disabling the
FirmwareUpdaterApp feature.

BUG=b:227267523
TEST=tast run dut arc.PowerIdlePerf.vm arc.AppLoadingPerf.vm
arc.CameraPerfExtraMetrics.vm

Change-Id: Iad403406ddacbc11c6e0458068e1fb1b05837e56