factory: Skip platforms not in factory and supply different params
Skip platforms that are not used in the factory for HWIDDatabase test,
thus updating the hardware dependency. Kitty is out of AUE and is not
used in test anymore, so removed it from the hardware dependency.

The test is separated by different parameters. The parameter added in
the CL is to automatically decline user prompt for not able to probe a
storage component, where this is an intended behavior on some models
and should run the hwid program with corresponding parameters.

BUG=b:223075358
TEST=tast run ${DUT} HWIDDatabase.default_settings &
     tast run ${DUT} HWIDDatabase.allow_no_storage

Change-Id: Ia792a9bfa4ab1aa52305404932caebb26da3e82c