tast: Change TapKeyJS to TapKey in ARC++ keyboard test.
Use TapKey instead of TapKeyJS so that the key event has valid
coordinates when tapping the switch keyboard key in this test. This
allows these coordinates to be used to check whether the target view has
updated, for multitouch purposes where multiple keys map be tapped at
the same time with some possibly changing the keyboard state.

BUG=b:225763175
TEST=tast run dut arc.ChromeVirtualKeyboard

Change-Id: Ic5797d4a037309f754646079d2d1204a192484d3