[ZipPerf] Improve time measurement
Polling in Tast tests is discouraged since it's quite expensive.
It's also inaccurate in our case because of the polling interval,
and the reliance on notifications which can be displayed with
a slight delay.

This change fixes the issue for mounting and zipping operations by
relying on recorded metrics instead of visual signals.

BUG=chromium:1315091
TEST=tast run $DUT filemanager.ZipPerf*

Change-Id: Ia29184973b91efeb19355698e91b2a804341c708
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3596597
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Jeremie Boulic <jboulic@chromium.org>
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>