Report more from chrome://webrtc-internals in ui.MeetCUJ.*
frameWidth, frameHeight, and framesPerSecond will be reported for each
active video stream in each ui.MeetCUJ test variant, except for 49p and
49p_vp8. chrome://webrtc-internals contains lots of info organized into
collapsible sections, and it can get extremely slow with 49 bots.

Caveat: If you sit and watch chrome://webrtc-internals, you will see
framesPerSecond fluctuate because it describes the number of frames that
were actually encoded/decoded in the past second (see [1]). In fact,
near the bottom of chrome://webrtc-internals, you will find graphs of
framesPerSecond over time (see the bottom of [2] for example). The tests
will simply read framesPerSecond at an arbitrary moment, because I think
that should be helpful enough. However you need to be aware of this so
you know how to interpret the results.

[1] https://www.w3.org/TR/webrtc-stats/
[2] https://drive.google.com/file/d/1cH5NAt6snbFJKQL-O37BudYjVYlrYdXv/view?usp=sharing

BUG=b:227355300
TEST=tast run <DUT> ui.MeetCUJ.*

Change-Id: I90f78a7a553e0f8f6692549e93f09f5504df2f50