RELAND: video/webrtc: collect GPU frequency stats
This CL relands the GPU collection code, which was reverted because
it returned error upon not finding the /sys/class/devfreq/ folder
(which happens on older Intel). This CL fixes that by handling
that case and not returning an error, following the example in
the sibling collect...Counters() functions.

PS0 contains the originally landed and reverted CL, subsequent
patches the fix.

Original commit description ------------------------------------------

This CL adds GPU frequency stats via reading of the sysfs
"trans_stat" file.

BUG=b:181562203
TEST=video.PlaybackPerf.h264_240p_30fps_hw on kevin, hana
(values are similar to what powertop shows, which was my
inspiration for this CL).

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3600865
Reviewed-by: Justin Green <greenjustin@google.com>
Tested-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Change-Id: Icd3075ceb8315d81427ca6b8f783e26fb3f2f6e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3625813