video: run only TenConcurrentDecoders on certain cases
RunAccelVideoPerfTest() runs a number of test cases; and since about
R100 it also runs a new TenConcurrentDecoders GTest case. This is run
on all devices, for both DecodeAccelPerf and DecodeAccelVDPerf, and
for all resolutions. This is unnecessary since the data intended to
be collected was only needed for Intel and AMD (VA-API platforms)
and mostly for DecodeAccelVDPerf (i.e. the direct VideoDecoder).
Moreover, this addition makes many small devices timeout for large
resolutions.

This CL adds a parameter to said function to only enable these extra
cases on demand.

BUG=b:230044499
TEST=video.DecodeAccelVDPerf.h264_1080p_60fps on hana

Change-Id: I5262efe50c93b2024ea133c43b0c3f8f0a5e3a30