Projector: Add Tast test for opening app and launching creation flow
This CL adds a Tast test that opens the Projector app and goes through
the creation flow. It dismisses the onboarding dialog, clicks the new
screencast button, opens the annotator, changes the marker color to
blue, draws a dot, clears the canvas, and stops the recording. It also
dismisses the tutorial videos and deletes all screencast items to
clean up for the next test.

This CL also adds a fixture so future Projector Tast tests can share
common setup code.

BUG=b:229631476,b:229632058,b:229631504,b:229634051,b:229634049,b:229633784,b:229634235,b:229631680,b:229634287,b:229632276,b:229634227,b:229632608
TEST=./repeat-test.sh -r 100 -- -buildbundle=cros localhost:2222 projector.CreationFlow

Change-Id: I8cb71f8ca7ce8a2beb6f8f02fdd2d55c65b3d97f