Projector: Expand SharedScreencast Tast test
After opening a share link and verifying that the shared screencast
rendered properly, this CL continues on to click the share button,
copy the shareable link to clipboard, and verify the clipboard data.

It also tests translating to French, searching the transcript,
navigating the transcript, skipping backwards and forwards, and
playing/pausing.

BUG=b:229633372,b:229633530,b:229633434,b:229633442,b:229633874,b:229633652,b:229633880,b:229632587
TEST=./repeat-test.sh -r 40 -- -buildbundle=cros localhost:2222 projector.*

Change-Id: Ib0b954653a88a285b4acbdfdee0059e70c000bf4