faft: Improve on BatteryCharging
In addition to running the host command 'power_supply_info', check
on its base files for verification on battery state, and use this
info in case that 'power_supply_info' did not update in time. Also,
log servo and dut connection type for debugging purposes, and wait
for DUT to reconnect when pd role is switched.

BUG=b:200305302
TEST=tast run <DUT> firmware.BatteryCharging

Change-Id: I5c804ac821e35c4180fe8aed1a0cd558aebad5d8