tast-tests: Increase logcat buffer size in enterprise.ARCProvisioning
When installation of apps fail, we need to share bugreport with the
Managed Play team and due to default buffer size of 256 Kb the
installation logs are flushed out by the time the report is collected.
In this change, we're increasing the buffer size to 10 MB so that the
extracted bugreport contains the necessary logs for app installation.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: Iac4fdf6dad2ec4a92eda0b2ddebb68b60f11b3a7